+++ src/tcn/py_ftn_interface/example/main.f90
@@ -1,0 +1,12 @@
+program test
+   use stub_interface_mod, only: python_function_f, data_t
+
+   implicit none
+
+   type(data_t) :: d
+   d = data_t(42.42, 24, .true.)
+   call python_function_f(d, 39)
+
+   print *, 'test'
+end program test
+