+++ sw_stack/discover/sles15/src/2024.03.00/build_1_on-login.sh
@@ -5,8 +5,8 @@ source ./basics.sh
 
 echo " === Make NDSL venv === "
 cd $DSLSW_INSTALL_DIR
-./python3/bin/python3 -m venv venv
+python3 -m venv venv
 source ./venv/bin/activate
 pip install --upgrade setuptools pip
 pip install -e $DSLSW_INSTALL_DIR/ndsl
-pip install mpi4py cffi
+pip install mpi4py cffi cupy-cuda12x